Fix Linux + Windows CI: dedup g_persp_profile, replace setenv, drop %zu
Three real cross-platform issues the macOS toolchain let slide:

* g_persp_profile was defined in both src/stubs.c and src/assets.c.
  Clang silently merged the tentative definitions; gcc + LTO on
  Linux choked with 'multiple definition'. Removed the stubs.c
  copy; assets.c stays canonical.

* src/main.c used POSIX setenv() to force SDL_VIDEODRIVER=dummy in
  headless mode — not in mingw's MSVCRT. Switched to SDL_setenv /
  SDL_getenv which is portable across every SDL platform.

* Three LOG_* call sites in src/main.c and src/save.c used %zu for
  size_t args. mingw GCC warns ('unknown conversion type z') against
  its default MSVCRT runtime. Cast to (unsigned long) / (unsigned)
  with %lu / %u so the format strings are portable everywhere.

Also aligned the g_walk_fld_* extern decls in bg_mask.c, walkability
.c, and render.c with the canonical 'int' definitions in game.c —
the mismatched uint16_t / int16_t externs triggered LTO type-mismatch
warnings ('code may be misoptimized'). Now all sites agree on int.

+++ src/actor/render.c
@@ -71,7 +71,7 @@ extern Entity *g_actor[2];
 
 extern void  *g_scene_bg_raw;
 extern uint32_t g_scene_bg_size;
-extern int16_t g_walk_fld_oy;
+extern int     g_walk_fld_oy;
 
 /* ---- depth-sort comparator ----------------------------------------- *
  *
